requirements.txt
banana-dev
scikit-image==0.18.3
Pillow==8.3.2